+++ Framework/Spell/Components/Harvest.cs
@@ -70,7 +70,7 @@ namespace ArsVenefici.Framework.Spell.Components
             }
         }
 
-        public override int ManaCost()
+        public override float ManaCost()
         {
             return 5;
         }
