+++ Framework/Spell/Components/Light.cs
@@ -92,7 +92,7 @@ namespace ArsVenefici.Framework.Spell.Components
             }
         }
 
-        public override int ManaCost()
+        public override float ManaCost()
         {
             return 7;
         }
